requirements.txt
langchain-couchbase==0.1.1
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.38.0
tqdm==4.66.5
google-generativeai==0.8.3
